pipeline/requirements.txt
uvicorn
gunicorn
fastapi
requests
gatenlp
pyyaml